backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
pycryptodome==3.19.0
numpy==1.24.3